reorganised method in controllers
