+++ src/Controller/MediaController.php
@@ -6,6 +6,7 @@ use App\Entity\Trick;
 use App\Entity\Media;
 use App\Form\Type\ImageType;
 use App\Form\Type\YoutubeType;
+use App\Form\Type\FeaturedType;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
@@ -133,7 +134,7 @@ class MediaController extends AbstractController
     }
 
     #[Route('/media/delete/{id}', name: 'delete_media')]
-    public function deleteMedia(int $id, ManagerRegistry $doctrine, Request $request): Response
+    public function deleteMedia(int $id, ManagerRegistry $doctrine): Response
     {
         $user = $this->getUser();
         if (empty($user)) {
@@ -158,6 +159,63 @@ class MediaController extends AbstractController
 
         return $this->redirectToRoute('trick_edit', ["slug" => $trick->getSlug()]);
 
+    }
+
+    #[Route('/trick/edit/featured/{slug}', name: 'trick_edit_featured')]
+    public function editFeatured(string $slug, ManagerRegistry $doctrine, Request $request): Response
+    {
+        $user = $this->getUser();
+        if (empty($user)) {
+            $this->addFlash('error', 'You have to be logged in to delete a trick');
+            return $this->redirectToRoute('home');
+        }
+        
+        $trick = $doctrine->getRepository(Trick::class)->findOneBy(['slug' => $slug]);
+        $media_list = $doctrine->getRepository(Media::class)->findBy([
+            'trick' => $trick,
+            'type' => 'image',
+        ]);
+        $form = $this->createForm(FeaturedType::class, $trick, ['data' => $media_list]);
+        $form->handleRequest($request);
+
+        if ($form->isSubmitted() && $form->isValid()) {
+
+            $featuredImage = $form->getData()["featuredImage"];            
+            $trick->setLastUpdate(new \DateTime());
+            $trick->setFeaturedImage($featuredImage);
+
+            $entityManager = $doctrine->getManager();
+            $entityManager->flush();
+
+            $this->addFlash('success', 'Trick successfully edited.');
+
+            return $this->redirectToRoute('trick_edit', ['slug' => $trick->getSlug()]);
+        }
+        
+        return $this->render('trick/edit_featured.html.twig', [
+            'trick' => $trick,
+            'form' => $form->createView(),
+        ]);
+    }
+
+    #[Route('/trick/edit/unlinkfeatured/{slug}', name: 'trick_unlink_featured')]
+    public function unlinkFeatured(string $slug, ManagerRegistry $doctrine): Response
+    {
+        $user = $this->getUser();
+        if (empty($user)) {
+            $this->addFlash('error', 'You have to be logged in to edit a trick');
+            return $this->redirectToRoute('home');
+        }
+        
+        $trick = $doctrine->getRepository(Trick::class)->findOneBy(['slug' => $slug]);
+        $trick->setFeaturedImage(null);
+
+        $entityManager = $doctrine->getManager();
+        $entityManager->flush();
+
+        $this->addFlash('success', 'Featured image unlinked successfully.');
+
+        return $this->redirectToRoute('trick_edit', ['slug' => $trick->getSlug()]);
     }
 
 }
